ov_core predeclare
