+++ ov_core/cmake/ROS1.cmake
@@ -50,6 +50,7 @@ list(APPEND LIBRARY_SOURCES
         src/track/TrackSIM.cpp
         src/types/Landmark.cpp
         src/feat/Feature.cpp
+        src/feat/FeatureDatabase.cpp
         src/feat/FeatureInitializer.cpp
         src/utils/print.cpp
 )
