+++ ov_core/src/track/TrackAruco.h
@@ -27,7 +27,6 @@
 #endif
 
 #include "TrackBase.h"
-#include "utils/print.h"
 
 namespace ov_core {
