+++ ov_core/src/track/TrackDescriptor.h
@@ -22,8 +22,6 @@
 #ifndef OV_CORE_TRACK_DESC_H
 #define OV_CORE_TRACK_DESC_H
 
-#include <opencv2/features2d.hpp>
-
 #include "TrackBase.h"
 
 namespace ov_core {
