+++ ov_core/src/track/TrackSIM.h
@@ -23,7 +23,6 @@
 #define OV_CORE_TRACK_SIM_H
 
 #include "TrackBase.h"
-#include "utils/print.h"
 
 namespace ov_core {
